+++ scvelo/core/tests/test_arithmetic.py
@@ -42,8 +42,8 @@ class TestClippedLog:
         a_logged = clipped_log(a, lb=lb, ub=ub, eps=eps)
 
         assert a_logged.shape == a.shape
-        assert (a_logged >= np.log(lb + eps)).all()
-        assert (a_logged <= np.log(ub - eps)).all()
+        assert (a_logged >= np.log(lb + eps, dtype="float32")).all()
+        assert (a_logged <= np.log(ub - eps, dtype="float32")).all()
 
     @given(
         a=arrays(
@@ -76,8 +76,8 @@ class TestClippedLog:
         a_logged = clipped_log(a, lb=lb, ub=ub, eps=eps)
 
         assert a_logged.shape == a.shape
-        assert (a_logged >= np.log(lb + eps)).all()
-        assert (a_logged <= np.log(ub - eps)).all()
+        assert (a_logged >= np.log(lb + eps, dtype="float32")).all()
+        assert (a_logged <= np.log(ub - eps, dtype="float32")).all()
 
 
 class TestInvert:
